refactored qschemas and qtargets
