+++ abjadext/__init__.py
@@ -1,1 +1,3 @@
-__path__ = __import__("pkgutil").extend_path(__path__, __name__)  # type: ignore
+__path__ = __import__("pkgutil").extend_path(  # type: ignore
+    __path__, __name__  # type: ignore
+)  # type: ignore
