+++ setup.py
@@ -29,7 +29,7 @@ if __name__ == "__main__":
         ],
         extras_require={
             "test": [
-                "black>=19.10b0",
+                "black>=20.8b1",
                 "flake8>=3.8.2",
                 "isort>=4.3.21",
                 "mypy>=0.770",
@@ -39,7 +39,7 @@ if __name__ == "__main__":
             ]
         },
         include_package_data=True,
-        install_requires=["abjad>=3.0"],
+        install_requires=["abjad>=3.1"],
         license="MIT",
         long_description=pathlib.Path("README.md").read_text(),
         keywords=", ".join(
